server: Modularize benefit schemas
