+++ server/polar/benefit/strategies/base/__init__.py
@@ -1,0 +1,18 @@
+from .properties import BenefitGrantProperties, BenefitProperties
+from .service import (
+    BenefitActionRequiredError,
+    BenefitPropertiesValidationError,
+    BenefitRetriableError,
+    BenefitServiceError,
+    BenefitServiceProtocol,
+)
+
+__all__ = [
+    "BenefitActionRequiredError",
+    "BenefitServiceProtocol",
+    "BenefitPropertiesValidationError",
+    "BenefitRetriableError",
+    "BenefitServiceError",
+    "BenefitProperties",
+    "BenefitGrantProperties",
+]
