+++ server/polar/benefit/strategies/base/properties.py
@@ -1,0 +1,9 @@
+from typing import TypedDict
+
+
+class BenefitProperties(TypedDict):
+    """Configurable properties for this benefit."""
+
+
+class BenefitGrantProperties(TypedDict):
+    """Benefit grant properties."""
