+++ server/polar/benefit/strategies/custom/properties.py
@@ -1,0 +1,8 @@
+from ..base.properties import BenefitGrantProperties, BenefitProperties
+
+
+class BenefitCustomProperties(BenefitProperties):
+    note: str | None
+
+
+class BenefitGrantCustomProperties(BenefitGrantProperties): ...
