+++ server/polar/benefit/strategies/github_repository/properties.py
@@ -1,0 +1,16 @@
+from typing import Literal
+
+from ..base.properties import BenefitGrantProperties, BenefitProperties
+
+
+class BenefitGitHubRepositoryProperties(BenefitProperties):
+    repository_owner: str
+    repository_name: str
+    permission: Literal["pull", "triage", "push", "maintain", "admin"]
+
+
+class BenefitGrantGitHubRepositoryProperties(BenefitGrantProperties, total=False):
+    account_id: str
+    repository_owner: str
+    repository_name: str
+    permission: Literal["pull", "triage", "push", "maintain", "admin"]
