+++ server/polar/benefit/strategies/github_repository/service.py
@@ -21,7 +21,7 @@ from polar.models.customer import CustomerOAuthPlatform
 from polar.posthog import posthog
 from polar.worker import compute_backoff
 
-from .base import (
+from ..base.service import (
     BenefitActionRequiredError,
     BenefitPropertiesValidationError,
     BenefitRetriableError,
